Adjust styles for filter modal
- Create two reusable hooks for removing code duplications
