+++ app/src/Component/Offering/FilterColumn.module.scss
@@ -5,6 +5,11 @@
   padding: 1rem;
 }
 
+.modal {
+  overflow-y: auto;
+  max-height: 75vh;
+}
+
 .hidden {
   display: none !important;
 }
@@ -14,6 +19,16 @@
 }
 
 .offeringFilterField {
+  margin: 5px 25px !important;
+}
+
+.offeringFilterField_modal {
+  @media only screen and (max-width: 1200px) {
+    margin: 5px 25px !important;
+  }
+}
+
+.offeringFilterDateField {
   margin: 5px 25px !important;
 }
 
@@ -27,5 +42,6 @@
 }
 
 .Filter_row {
+  margin: 4px 0;
   align-items: center;
 }
