+++ app/src/Component/Offering/FilterColumnModal.tsx
@@ -1,21 +1,42 @@
-import React, { useState, useEffect } from "react"
+import React, { useState } from "react"
 import { Col, Row, Checkbox, Input, Select, Button, Typography, DatePicker } from "antd"
 import { CloseOutlined } from "@ant-design/icons"
 import moment from "moment"
 import { CheckboxChangeEvent } from "antd/lib/checkbox"
 import styles from "~/Component/Offering/FilterColumn.module.scss"
-import {
-  getOfferingStatusTypes,
-  getOrganizations,
-  getOfferingTypes,
-  getSectionTypes,
-  getTagTypes
-} from "~/ApiServices/Service/RefLookupService"
-import { getUsersByRole } from "~/ApiServices/Service/HRUserService"
+
+import { ColProps } from "antd/lib/col"
+import { useFilterData } from "~/Component/Offering/offeringUtils"
 
 const { Option } = Select
 const { Title } = Typography
 
+const layout = {
+  label: {
+    md: 24,
+    lg: 24,
+    xl: 10,
+    xxl: 10,
+    sm: 24,
+    xs: 24
+  },
+  input: {
+    md: 20,
+    lg: 20,
+    xl: 14,
+    xxl: 14,
+    sm: 20,
+    xs: 20
+  }
+}
+
+function LabelCol(props: ColProps) {
+  return <Col {...layout.label} {...props} />
+}
+
+function InputCol(props: ColProps) {
+  return <Col {...layout.input} {...props} />
+}
 export interface IFilterValues {
   OfferingCode: string
   OfferingName: string
@@ -59,51 +80,7 @@ interface IFilterColumnProps {
 }
 
 export function FilterColumn(props: IFilterColumnProps) {
-  const [offeringStatusTypes, setOfferingStatusTypes] = useState<Array<any>>([])
-  const [tagTypes, setTagTypes] = useState<Array<any>>([])
-  const [offeringTypes, setOfferingTypes] = useState<Array<any>>([])
-  const [sectonTypes, setSectionTypes] = useState<Array<any>>([])
-  const [organizations, setOrganizations] = useState<Array<any>>([])
-  const [users, setUsers] = useState<Array<any>>([])
-
-  useEffect(() => {
-    ;(async () => {
-      const response = await getOfferingStatusTypes()
-      if (response && response.data && Array.isArray(response.data)) {
-        setOfferingStatusTypes(response.data)
-      }
-    })()
-    ;(async () => {
-      const response = await getOrganizations()
-      if (response && response.data) {
-        setOrganizations(response.data)
-      }
-    })()
-    ;(async () => {
-      const response = await getOfferingTypes()
-      if (response && response.data) {
-        setOfferingTypes(response.data)
-      }
-    })()
-    ;(async () => {
-      const response = await getUsersByRole({ Role: "coordinator" })
-      if (response && response.data) {
-        setUsers(response.data)
-      }
-    })()
-    ;(async () => {
-      const response = await getSectionTypes()
-      if (response && response.success) {
-        setSectionTypes(response.data)
-      }
-    })()
-    ;(async () => {
-      const response = await getTagTypes()
-      if (response && response.success) {
-        setTagTypes(response.data)
-      }
-    })()
-  }, [])
+  const [offeringStatusTypes, tagTypes, offeringTypes, sectonTypes, organizations, users] = useFilterData()
 
   const { visible, toggleVisiibility, data } = props
   const [filterData, updateFilterData] = useState<IFilterValues>(data)
@@ -317,229 +294,284 @@ export function FilterColumn(props: IFilterColumnProps) {
 
   return (
     <Col className={visible ? `gutter-row ${styles.offeringFilter}` : styles.hidden} xs={24} sm={24} md={12}>
-      <Row>
-        <Col span={12}>
-          <Title level={4}>Offering Filter</Title>
-        </Col>
-        <Col span={12} className={styles.padding5px}>
-          <span onClick={toggleVisiibility}>
-            <CloseOutlined style={{ fontSize: "20px", color: "black", float: "right" }} />
-          </span>
-        </Col>
-      </Row>
-      <Row className={styles.Filter_row}>
-        <Checkbox onChange={toggleOfferingCodeBLock}>Offering Code</Checkbox>
-        <Row className={showOfferingCodeBlock ? styles.offeringFilterField : styles.hidden}>
-          <Input
-            name="OfferingCode"
-            defaultValue=""
-            value={filterData.OfferingCode === "*" ? "" : filterData.OfferingCode}
-            onChange={handleInputChange}
-          />
+      <Col className={styles.modal}>
+        <Row>
+          <Col span={12}>
+            <Title level={4}>Offering Filter</Title>
+          </Col>
+          <Col span={12} className={styles.padding5px}>
+            <span onClick={toggleVisiibility}>
+              <CloseOutlined style={{ fontSize: "20px", color: "black", float: "right" }} />
+            </span>
+          </Col>
         </Row>
-      </Row>
-      <Row className={styles.Filter_row}>
-        <Checkbox onChange={toggleOfferingNameBLock}>Offering Name</Checkbox>
-        <Row className={showOfferingNameBlock ? styles.offeringFilterField : styles.hidden}>
-          <Input
-            name="OfferingName"
-            defaultValue=""
-            value={filterData.OfferingName === "*" ? "" : filterData.OfferingName}
-            onChange={handleInputChange}
-          />
-        </Row>
-      </Row>
-      <Row>
-        <Checkbox onChange={toggleCreationDateBLock}>Creation Date</Checkbox>
-        <Row className={showCreationDateBlock ? styles.offeringFilterField : styles.hidden}>
-          <Col span={24}>From</Col>
-          <DatePicker allowClear value={fromCreationDate} onChange={handleFromCreationDateChange} format={dateFormat} />
-          <Col span={24}>To</Col>
-          <DatePicker allowClear value={toCreationDate} onChange={handleToCreationDateChange} format={dateFormat} />
-        </Row>
-      </Row>
-      <Row>
-        <Checkbox onChange={toggleTerminationDateBLock}>Termination Date</Checkbox>
-        <Row className={showTerminationDateBlock ? styles.offeringFilterField : styles.hidden}>
-          <Col span={24}>From</Col>
-          <DatePicker
-            allowClear
-            value={fromTerminationDate}
-            onChange={handleFromTerminationDateChange}
-            format={dateFormat}
-          />
-          <Col span={24}>To</Col>
-          <DatePicker
-            allowClear
-            value={toTerminationDate}
-            onChange={handleToTerminationDateChange}
-            format={dateFormat}
-          />
-        </Row>
-      </Row>
-      <Row className={styles.Filter_row}>
-        <Checkbox onChange={toggleIsQuickAdmitBLock}>Is QuickAdmit</Checkbox>
-        <Row className={showIsQuickAdmitBlock ? styles.offeringFilterField : styles.hidden}>
-          <Select style={{ width: 200 }} value={filterData.IsQuickAdmit} onChange={onChangeSelect("IsQuickAdmit")}>
-            <Option value="true">Yes</Option>
-            <Option value="false">No</Option>
-          </Select>
-        </Row>
-      </Row>
-      {offeringStatusTypes.length > 0 && (
         <Row className={styles.Filter_row}>
-          <Checkbox onChange={toggleOfferingStatusBLock}>Offering Status</Checkbox>
-          <Row className={showOfferingStatusBlock ? styles.offeringFilterField : styles.hidden}>
-            <Select style={{ width: 200 }} value={filterData.StatusID} onChange={onChangeSelect("StatusID")}>
-              {offeringStatusTypes.map((x) => {
-                return (
-                  <Select.Option key={x.StatusID} value={x.StatusID}>
-                    {x.Name}
-                  </Select.Option>
-                )
-              })}
-            </Select>
+          <LabelCol>
+            <Checkbox onChange={toggleOfferingCodeBLock}>Offering Code</Checkbox>
+          </LabelCol>
+          <InputCol className={showOfferingCodeBlock ? styles.offeringFilterField_modal : styles.hidden}>
+            <Input
+              name="OfferingCode"
+              defaultValue=""
+              value={filterData.OfferingCode === "*" ? "" : filterData.OfferingCode}
+              onChange={handleInputChange}
+            />
+          </InputCol>
+        </Row>
+        <Row className={styles.Filter_row}>
+          <LabelCol>
+            <Checkbox onChange={toggleOfferingNameBLock}>Offering Name</Checkbox>
+          </LabelCol>
+          <InputCol className={showOfferingNameBlock ? styles.offeringFilterField_modal : styles.hidden}>
+            <Input
+              name="OfferingName"
+              defaultValue=""
+              value={filterData.OfferingName === "*" ? "" : filterData.OfferingName}
+              onChange={handleInputChange}
+            />
+          </InputCol>
+        </Row>
+        <Row className={styles.Filter_row}>
+          <LabelCol>
+            <Checkbox onChange={toggleCreationDateBLock}>Creation Date</Checkbox>
+          </LabelCol>
+          <Row className={showCreationDateBlock ? styles.offeringFilterDateField : styles.hidden}>
+            <Col span={11}>
+              From
+              <DatePicker
+                allowClear
+                value={fromCreationDate}
+                onChange={handleFromCreationDateChange}
+                format={dateFormat}
+              />
+            </Col>
+
+            <Col span={11} offset={2}>
+              To
+              <DatePicker allowClear value={toCreationDate} onChange={handleToCreationDateChange} format={dateFormat} />
+            </Col>
           </Row>
         </Row>
-      )}
-      {offeringTypes.length > 0 && (
         <Row className={styles.Filter_row}>
-          <Checkbox onChange={toggleOfferingTypeBLock}>Offering Type</Checkbox>
-          <Row className={showOfferingTypeBlock ? styles.offeringFilterField : styles.hidden}>
+          <LabelCol>
+            <Checkbox onChange={toggleTerminationDateBLock}>Termination Date</Checkbox>
+          </LabelCol>
+
+          <Row className={showTerminationDateBlock ? styles.offeringFilterDateField : styles.hidden}>
+            <Col span={11}>
+              From
+              <DatePicker
+                allowClear
+                value={fromTerminationDate}
+                onChange={handleFromTerminationDateChange}
+                format={dateFormat}
+              />
+            </Col>
+            <Col span={11} offset={2}>
+              To
+              <DatePicker
+                allowClear
+                value={toTerminationDate}
+                onChange={handleToTerminationDateChange}
+                format={dateFormat}
+              />
+            </Col>
+          </Row>
+        </Row>
+        <Row className={styles.Filter_row}>
+          <LabelCol>
+            <Checkbox onChange={toggleIsQuickAdmitBLock}>Is QuickAdmit</Checkbox>
+          </LabelCol>
+          <InputCol className={showIsQuickAdmitBlock ? styles.offeringFilterField_modal : styles.hidden}>
+            <Select style={{ width: 200 }} value={filterData.IsQuickAdmit} onChange={onChangeSelect("IsQuickAdmit")}>
+              <Option value="true">Yes</Option>
+              <Option value="false">No</Option>
+            </Select>
+          </InputCol>
+        </Row>
+        {offeringStatusTypes.length > 0 && (
+          <Row className={styles.Filter_row}>
+            <LabelCol>
+              <Checkbox onChange={toggleOfferingStatusBLock}>Offering Status</Checkbox>
+            </LabelCol>
+            <InputCol className={showOfferingStatusBlock ? styles.offeringFilterField_modal : styles.hidden}>
+              <Select style={{ width: 200 }} value={filterData.StatusID} onChange={onChangeSelect("StatusID")}>
+                {offeringStatusTypes.map((x) => {
+                  return (
+                    <Select.Option key={x.StatusID} value={x.StatusID}>
+                      {x.Name}
+                    </Select.Option>
+                  )
+                })}
+              </Select>
+            </InputCol>
+          </Row>
+        )}
+        {offeringTypes.length > 0 && (
+          <Row className={styles.Filter_row}>
+            <LabelCol>
+              <Checkbox onChange={toggleOfferingTypeBLock}>Offering Type</Checkbox>
+            </LabelCol>
+            <InputCol className={showOfferingTypeBlock ? styles.offeringFilterField_modal : styles.hidden}>
+              <Select
+                style={{ width: 200 }}
+                value={filterData.OfferingTypeID}
+                onChange={onChangeSelect("OfferingTypeID")}
+              >
+                {offeringTypes.map((x) => {
+                  return (
+                    <Select.Option key={x.OfferingTypeID} value={x.OfferingTypeID}>
+                      {x.Name}
+                    </Select.Option>
+                  )
+                })}
+              </Select>
+            </InputCol>
+          </Row>
+        )}
+        {organizations.length > 0 && (
+          <Row className={styles.Filter_row}>
+            <LabelCol>
+              <Checkbox onChange={toggleDepartmentBLock}>Department</Checkbox>
+            </LabelCol>
+            <InputCol className={showDepartmentBlock ? styles.offeringFilterField_modal : styles.hidden}>
+              <Select style={{ width: 200 }} onChange={onChangeSelect("OrganizationID")}>
+                {organizations.map((x) => {
+                  return (
+                    <Select.Option key={x.OrganizationTypeID} value={x.OrganizationTypeID}>
+                      {x.Name}
+                    </Select.Option>
+                  )
+                })}
+              </Select>
+            </InputCol>
+          </Row>
+        )}
+        {users.length > 0 && (
+          <Row className={styles.Filter_row}>
+            <LabelCol>
+              <Checkbox onChange={toggleCoordinatorBLock}>Coordinator</Checkbox>
+            </LabelCol>
+            <InputCol className={showCoordinatorBlock ? styles.offeringFilterField_modal : styles.hidden}>
+              <Select style={{ width: 200 }} value={filterData.Coordinator} onChange={onChangeSelect("Coordinator")}>
+                {users.map((x) => {
+                  return (
+                    <Select.Option key={x.UserLogin} value={x.UserLogin}>
+                      {x.FormattedName}
+                    </Select.Option>
+                  )
+                })}
+              </Select>
+            </InputCol>
+          </Row>
+        )}
+        {sectonTypes.length > 0 && (
+          <Row className={styles.Filter_row}>
+            <LabelCol>
+              <Checkbox onChange={toggleSectionTypeBLock}>Section Type</Checkbox>
+            </LabelCol>
+            <InputCol className={showSectionTypeBlock ? styles.offeringFilterField_modal : styles.hidden}>
+              <Select
+                style={{ width: 200 }}
+                value={filterData.SectionTypeID}
+                onChange={onChangeSelect("SectionTypeID")}
+              >
+                {sectonTypes.map((x) => {
+                  return (
+                    <Select.Option key={x.SectionTypeID} value={x.SectionTypeID}>
+                      {x.SectionTypeName}
+                    </Select.Option>
+                  )
+                })}
+              </Select>
+            </InputCol>
+          </Row>
+        )}
+        <Row className={styles.Filter_row}>
+          <LabelCol>
+            <Checkbox onChange={toggleIsSearchTagHierarchyBLock}>Is Search Tag Hierarchy</Checkbox>
+          </LabelCol>
+          <InputCol className={showIsSearchTagHierarchyBlock ? styles.offeringFilterField_modal : styles.hidden}>
             <Select
               style={{ width: 200 }}
-              value={filterData.OfferingTypeID}
-              onChange={onChangeSelect("OfferingTypeID")}
+              value={filterData.IsSearchTagHierarchy}
+              onChange={onChangeSelect("IsSearchTagHierarchy")}
             >
-              {offeringTypes.map((x) => {
-                return (
-                  <Select.Option key={x.OfferingTypeID} value={x.OfferingTypeID}>
-                    {x.Name}
-                  </Select.Option>
-                )
-              })}
+              <Option value="true">Yes</Option>
+              <Option value="false">No</Option>
             </Select>
-          </Row>
+          </InputCol>
         </Row>
-      )}
-      {organizations.length > 0 && (
+        {tagTypes.length > 0 && (
+          <Row className={styles.Filter_row}>
+            <LabelCol>
+              <Checkbox onChange={toggleTagBLock}>Tag Type</Checkbox>
+            </LabelCol>
+            <InputCol className={showTagTypeBlock ? styles.offeringFilterField_modal : styles.hidden}>
+              <Select style={{ width: 200 }} value={filterData.TagTypeID} onChange={onChangeSelect("TagTypeID")}>
+                {tagTypes.map((x) => {
+                  return (
+                    <Select.Option key={x.ID} value={x.ID}>
+                      {x.Name}
+                    </Select.Option>
+                  )
+                })}
+              </Select>
+            </InputCol>
+          </Row>
+        )}
         <Row className={styles.Filter_row}>
-          <Checkbox onChange={toggleDepartmentBLock}>Department</Checkbox>
-          <Row className={showDepartmentBlock ? styles.offeringFilterField : styles.hidden}>
-            <Select style={{ width: 200 }} onChange={onChangeSelect("OrganizationID")}>
-              {organizations.map((x) => {
-                return (
-                  <Select.Option key={x.OrganizationTypeID} value={x.OrganizationTypeID}>
-                    {x.Name}
-                  </Select.Option>
-                )
-              })}
-            </Select>
+          <LabelCol>
+            <Checkbox onChange={toggleTagNameBLock}>Tag</Checkbox>
+          </LabelCol>
+          <InputCol className={showTagNameBlock ? styles.offeringFilterField_modal : styles.hidden}>
+            <Input
+              name="TagName"
+              defaultValue=""
+              value={filterData.TagName === "*" ? "" : filterData.TagName}
+              onChange={handleInputChange}
+            />
+          </InputCol>
+        </Row>
+        <Row>
+          <LabelCol>
+            <Checkbox onChange={toggleFinalEnrollmentBLock}>Final Enrollment Date</Checkbox>
+          </LabelCol>
+          <Row className={showFinalEnrollmentBlock ? styles.offeringFilterDateField : styles.hidden}>
+            <Col span={11}>
+              From
+              <DatePicker
+                allowClear
+                value={fromFinalEnrollmentDate}
+                onChange={handleFromFinalEnrollmentDateChange}
+                format={dateFormat}
+              />
+            </Col>
+            <Col span={11} offset={2}>
+              To
+              <DatePicker
+                allowClear
+                value={toFinalEnrollmentDate}
+                onChange={handleToFinalEnrollmentDateChange}
+                format={dateFormat}
+              />
+            </Col>
           </Row>
         </Row>
-      )}
-      {users.length > 0 && (
         <Row className={styles.Filter_row}>
-          <Checkbox onChange={toggleCoordinatorBLock}>Coordinator</Checkbox>
-          <Row className={showCoordinatorBlock ? styles.offeringFilterField : styles.hidden}>
-            <Select style={{ width: 200 }} value={filterData.Coordinator} onChange={onChangeSelect("Coordinator")}>
-              {users.map((x) => {
-                return (
-                  <Select.Option key={x.UserLogin} value={x.UserLogin}>
-                    {x.FormattedName}
-                  </Select.Option>
-                )
-              })}
-            </Select>
-          </Row>
+          <LabelCol>
+            <Checkbox onChange={toggleOfferingNearCapacityBLock}>Capacity Util</Checkbox>
+          </LabelCol>
+          <InputCol className={showOfferingNearCapacityBlock ? styles.offeringFilterField_modal : styles.hidden}>
+            <Input
+              name="OfferingNearCapacity"
+              defaultValue=""
+              value={filterData.OfferingNearCapacity === "*" ? "" : filterData.OfferingNearCapacity}
+              onChange={handleInputChange}
+            />
+          </InputCol>
         </Row>
-      )}
-      {sectonTypes.length > 0 && (
-        <Row className={styles.Filter_row}>
-          <Checkbox onChange={toggleSectionTypeBLock}>Section Type</Checkbox>
-          <Row className={showSectionTypeBlock ? styles.offeringFilterField : styles.hidden}>
-            <Select style={{ width: 200 }} value={filterData.SectionTypeID} onChange={onChangeSelect("SectionTypeID")}>
-              {sectonTypes.map((x) => {
-                return (
-                  <Select.Option key={x.SectionTypeID} value={x.SectionTypeID}>
-                    {x.SectionTypeName}
-                  </Select.Option>
-                )
-              })}
-            </Select>
-          </Row>
-        </Row>
-      )}
-      <Row className={styles.Filter_row}>
-        <Checkbox onChange={toggleIsSearchTagHierarchyBLock}>Is Search Tag Hierarchy</Checkbox>
-        <Row className={showIsSearchTagHierarchyBlock ? styles.offeringFilterField : styles.hidden}>
-          <Select
-            style={{ width: 200 }}
-            value={filterData.IsSearchTagHierarchy}
-            onChange={onChangeSelect("IsSearchTagHierarchy")}
-          >
-            <Option value="true">Yes</Option>
-            <Option value="false">No</Option>
-          </Select>
-        </Row>
-      </Row>
-      {tagTypes.length > 0 && (
-        <Row className={styles.Filter_row}>
-          <Checkbox onChange={toggleTagBLock}>Tag Type</Checkbox>
-          <Row className={showTagTypeBlock ? styles.offeringFilterField : styles.hidden}>
-            <Select style={{ width: 200 }} value={filterData.TagTypeID} onChange={onChangeSelect("TagTypeID")}>
-              {tagTypes.map((x) => {
-                return (
-                  <Select.Option key={x.ID} value={x.ID}>
-                    {x.Name}
-                  </Select.Option>
-                )
-              })}
-            </Select>
-          </Row>
-        </Row>
-      )}
-      <Row className={styles.Filter_row}>
-        <Checkbox onChange={toggleTagNameBLock}>Tag</Checkbox>
-        <Row className={showTagNameBlock ? styles.offeringFilterField : styles.hidden}>
-          <Input
-            name="TagName"
-            defaultValue=""
-            value={filterData.TagName === "*" ? "" : filterData.TagName}
-            onChange={handleInputChange}
-          />
-        </Row>
-      </Row>
-      <Row>
-        <Checkbox onChange={toggleFinalEnrollmentBLock}>Final Enrollment Date</Checkbox>
-        <Row className={showFinalEnrollmentBlock ? styles.offeringFilterField : styles.hidden}>
-          <Col span={24}>From</Col>
-          <DatePicker
-            allowClear
-            value={fromFinalEnrollmentDate}
-            onChange={handleFromFinalEnrollmentDateChange}
-            format={dateFormat}
-          />
-          <Col span={24}>To</Col>
-          <DatePicker
-            allowClear
-            value={toFinalEnrollmentDate}
-            onChange={handleToFinalEnrollmentDateChange}
-            format={dateFormat}
-          />
-        </Row>
-      </Row>
-      <Row className={styles.Filter_row}>
-        <Checkbox onChange={toggleOfferingNearCapacityBLock}>Capacity Util</Checkbox>
-        <Row className={showOfferingNearCapacityBlock ? styles.offeringFilterField : styles.hidden}>
-          <Input
-            name="OfferingNearCapacity"
-            defaultValue=""
-            value={filterData.OfferingNearCapacity === "*" ? "" : filterData.OfferingNearCapacity}
-            onChange={handleInputChange}
-          />
-        </Row>
-      </Row>
+      </Col>
       <Row className={styles.floatRight}>
         <Button
           type="primary"
