+++ app/src/pages/Offering/index.tsx
@@ -9,6 +9,7 @@ import { searchOffering } from "~/ApiServices/Service/OfferingService"
 import { RouteComponentProps, Link } from "react-router-dom"
 import styles from "~/pages/Offering/Offering.module.scss"
 import { REFRESH_OFFERING_PAGE, eventBus } from "~/utils/EventBus"
+import { useOfferings } from "~/Component/Offering/offeringUtils"
 
 const { useState, useEffect } = React
 const { Title } = Typography
@@ -116,9 +117,6 @@ function expandableRowRender(data: any, display: boolean) {
 function OfferingPage(props: RouteComponentProps) {
   const [filterData, updateFilterData] = useState<IFilterValues>(INITIAL_FILTER_DATA)
   const [showFilter, setFilterVisiblity] = useState<boolean>(false)
-  const [loading, setLoading] = useState<boolean>(false)
-  const [offeringItems, setOfferingItems] = useState<Array<any>>([])
-
   const [filterCount, setFilterCount] = useState<number>(0)
 
   const { useBreakpoint } = Grid
@@ -195,58 +193,7 @@ function OfferingPage(props: RouteComponentProps) {
     }
   ]
 
-  useEffect(() => {
-    const loadOfferings = async function () {
-      setLoading(true)
-
-      console.log("Filter data: ", filterData)
-
-      const params: { [key: string]: any } = {}
-      params["OfferingCode"] = filterData.OfferingCode !== "" ? filterData.OfferingCode : "*"
-      params["OfferingName"] = filterData.OfferingName !== "" ? filterData.OfferingName : undefined
-      params["ToCreationDate"] = filterData.ToCreationDate !== "" ? filterData.ToCreationDate : undefined
-      params["FromCreationDate"] = filterData.FromCreationDate !== "" ? filterData.FromCreationDate : undefined
-      params["ToTerminationDate"] = filterData.ToTerminationDate !== "" ? filterData.ToTerminationDate : undefined
-      params["FromTerminationDate"] = filterData.FromTerminationDate !== "" ? filterData.FromTerminationDate : undefined
-      params["StatusID"] = filterData.StatusID !== "" ? Number(filterData.StatusID) : undefined
-      params["Coordinator"] = filterData.Coordinator !== "" ? filterData.Coordinator : undefined
-      params["OrganizationID"] = filterData.OrganizationID !== "" ? Number(filterData.OrganizationID) : undefined
-      params["OfferingTypeID"] = filterData.OfferingTypeID !== "" ? Number(filterData.OfferingTypeID) : undefined
-      params["SectionTypeID"] = filterData.SectionTypeID !== "" ? Number(filterData.SectionTypeID) : undefined
-      params["InstructorID"] = filterData.InstructorID !== "" ? Number(filterData.InstructorID) : undefined
-      params["ShowProgramOffering"] = filterData.ShowProgramOffering !== "" ? filterData.ShowProgramOffering : undefined
-      params["OfferingNearCapacity"] =
-        filterData.OfferingNearCapacity !== "" ? filterData.OfferingNearCapacity : undefined
-      params["IsQuickAdmit"] = filterData.IsQuickAdmit !== "" ? Boolean(filterData.IsQuickAdmit) : undefined
-      params["IsSearchTagHierarchy"] =
-        filterData.IsSearchTagHierarchy !== "" ? Boolean(filterData.IsSearchTagHierarchy) : undefined
-      params["TagName"] = filterData.TagName !== "" ? filterData.TagName : undefined
-      params["TagTypeID"] = filterData.TagTypeID !== "" ? filterData.TagTypeID : undefined
-      params["ToFinalEnrollmentDate"] =
-        filterData.ToFinalEnrollmentDate !== "" ? filterData.ToFinalEnrollmentDate : undefined
-      params["FromFinalEnrollmentDate"] =
-        filterData.FromFinalEnrollmentDate !== "" ? filterData.FromFinalEnrollmentDate : undefined
-
-      const objectKeys = Object.keys(params)
-      objectKeys.forEach((key) => {
-        if (!Boolean(params[key]) && typeof params[key] !== "number") {
-          delete params[key]
-        }
-      })
-
-      const result = await searchOffering(params)
-
-      if (result && result.success) {
-        setOfferingItems(result.data)
-      }
-      setLoading(false)
-    }
-    eventBus.subscribe(REFRESH_OFFERING_PAGE, loadOfferings)
-    eventBus.publish(REFRESH_OFFERING_PAGE)
-    return () => {
-      eventBus.unsubscribe(REFRESH_OFFERING_PAGE)
-    }
-  }, [filterData])
+  const [loading, offeringItems] = useOfferings(filterData)
 
   const toggleFilter = () => {
     setFilterVisiblity(!showFilter)
@@ -272,7 +219,7 @@ function OfferingPage(props: RouteComponentProps) {
           className={`gutter-row ${styles.offeringDetails}`}
           xs={24}
           sm={24}
-          md={{ span: showFilter ? 18 : 24, offset: showFilter ? 1 : 0 }}
+          md={{ span: showFilter ? 16 : 24, offset: showFilter ? 1 : 0 }}
         >
           <Table
             columns={columns}
